update timeutil test

+++ time/timeutil/timeutil_test.go
@@ -39,8 +39,8 @@ func TestDt8ForString(t *testing.T) {
 }
 
 var monthToQuarterTests = []struct {
-	v    int
-	want int
+	v    uint8
+	want uint8
 }{
 	{1, 1}, {2, 1}, {3, 1},
 	{4, 2}, {5, 2}, {6, 2},
@@ -57,8 +57,8 @@ func TestMonthToQuarter(t *testing.T) {
 }
 
 var quarterToMonthTests = []struct {
-	v    int
-	want int
+	v    uint8
+	want uint8
 }{
 	{1, 1}, {2, 4}, {3, 7}, {4, 10}}
